Render the entire map (#36)
* Add rendering for unbreakable blocks

* Add rendering for floor

* Clean Map.cpp code duplication

* Add dynamic camera and light positioning

+++ Engine/include/ECS/Components/Model3D.hpp
@@ -23,7 +23,7 @@ protected:
     irr::scene::IAnimatedMeshSceneNode *_node{nullptr};
     std::string _model{};
     std::string _texture{};
-    uint8_t _textureLayer{0};
+    uint32_t _textureLayer{0};
 
 public:
     Model3D(const Math::Vec3f &pos, const decltype(_model) &model);
